requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
easyocr==1.7.1
scikit-learn==1.3.0
tqdm==4.65.0
torch==2.1.0
torchvision==0.16.0
